wip: split out the harvest test from responses, changed reporter so i can see why these are failing
