+++ test/integration/newrelic-response-handling.tap.js
@@ -35,7 +35,7 @@ const endpointDataChecks = {
     return agent.spanEventAggregator.length > 0
   },
   custom_event_data: function hasCustomEventData(agent) {
-    // TODO... prob don't ned to grrab events
+    // TODO... prob don't need to grab events
     return agent.customEventAggregator.length > 0
   },
   sql_trace_data: function hasSqlTraceData(agent) {
@@ -44,123 +44,11 @@ const endpointDataChecks = {
 }
 
 tap.test('New Relic response code handling', (t) => {
-  t.autoend()
-  // t.plan(testCases.length)
-
-  t.before(() => {
-    nock.disableNetConnect()
-  })
-
-  t.teardown(() => {
-    if (!nock.isDone()) {
-      // eslint-disable-next-line no-console
-      console.error('Cleaning pending mocks: %j', nock.pendingMocks())
-      nock.cleanAll()
-    }
-
-    nock.enableNetConnect()
-  })
+  t.plan(testCases.length)
 
   testCases.forEach((testCase) => {
     const testName = `Status code: ${testCase.code}`
     t.test(testName, createStatusCodeTest(testCase))
-  })
-})
-
-/**
- * This test asserts that when the agent re-connects it pulls the harvest limits from the original
- * config max_samples_stored for every piece of data. This is done so on restart loops we aren't using
- * the new harvest limit value from server that has been down sampled.  It could result in harvest limits
- * being 0 if enough restarts occur.
- */
-tap.test('Connect calls re-generate harvest limits from original config values', (t) => {
-  let agent
-  let serverHarvest
-
-  t.before(() => {
-    serverHarvest = {
-      event_harvest_config: {
-        report_period_ms: 100,
-        harvest_limits: {
-          analytic_event_data: 10,
-          custom_event_data: 10,
-          error_event_data: 1,
-          span_event_data: 10,
-          log_event_data: 10
-        }
-      }
-    }
-    nock.disableNetConnect()
-    nockRequest('preconnect').reply(200, { return_value: TEST_DOMAIN })
-    nockRequest('connect').reply(200, { return_value: { agent_run_id: RUN_ID, ...serverHarvest } })
-    nockRequest('agent_settings', RUN_ID).reply(200, { return_value: [] })
-    nockRequest('metric_data', RUN_ID).reply(409, { return_value: [] })
-    nockRequest('preconnect').reply(200, { return_value: TEST_DOMAIN })
-    nockRequest('connect').reply(200, { return_value: { agent_run_id: RUN_ID, ...serverHarvest } })
-    nockRequest('agent_settings', RUN_ID).reply(200, { return_value: [] })
-    agent = helper.loadMockedAgent({
-      license_key: 'license key here',
-      host: TEST_DOMAIN,
-      application_logging: {
-        enabled: true
-      }
-    })
-  })
-
-  t.teardown(() => {
-    helper.unloadAgent(agent)
-    if (!nock.isDone()) {
-      // eslint-disable-next-line no-console
-      console.error('Cleaning pending mocks: %j', nock.pendingMocks())
-      nock.cleanAll()
-    }
-
-    nock.enableNetConnect()
-  })
-
-  const originalConfig = Object.assign({}, agent.config)
-  agent.config.no_immediate_harvest = true
-  sinon.spy(agent.collector, '_connect')
-
-  /**
-   * This flow starts agent which pre-connects, connects and gets agent settings.
-   * Then we call send metrics and since the metrics collector endpoint is responding
-   * with 409 it will issue a restart and make another pre-connect, connect and agent
-   * settings call.
-   */
-  agent.start((err) => {
-    t.error(err)
-    const config = agent.config
-    t.same(
-      config.event_harvest_config,
-      serverHarvest.event_harvest_config,
-      'config should have been updated from server'
-    )
-    agent.metrics.once('finished metric_data data send.', function onMetricsFinished() {
-      const connectCalls = agent.collector._connect.args
-      t.same(
-        config.event_harvest_config,
-        serverHarvest.event_harvest_config,
-        'config should have been updated from server after reconnect'
-      )
-      t.equal(connectCalls.length, 2, 'should have reconnected once')
-      connectCalls.forEach((call) => {
-        const factsConfig = call[0][0]
-        t.not(
-          factsConfig.event_harvest_config.harvest_limits,
-          config.event_harvest_config.harvest_limits,
-          'facts harvest config should not be same as new harvest config'
-        )
-        t.same(
-          factsConfig.event_harvest_config.harvest_limits,
-          originalConfig.event_harvest_config.harvest_limits,
-          'connect should send up original harvest limits'
-        )
-      })
-      t.end()
-    })
-
-    agent.metrics.send()
   })
 })
 
@@ -177,9 +65,8 @@ function createStatusCodeTest(testCase) {
 
     let agent = null
 
-    statusCodeTest.autoend()
     statusCodeTest.beforeEach(async () => {
-      // nock.disableNetConnect()
+      nock.disableNetConnect()
 
       testClock = sinon.useFakeTimers({
         toFake: ['setTimeout', 'setInterval', 'Date', 'clearInterval']
@@ -223,20 +110,19 @@ function createStatusCodeTest(testCase) {
       restartEndpoints = null
       shutdown = null
 
-      /* if (!nock.isDone()) {
+      if (!nock.isDone()) {
         // eslint-disable-next-line no-console
         console.error('Cleaning pending mocks: %j', nock.pendingMocks())
         nock.cleanAll()
       }
 
       nock.enableNetConnect()
-      */
     })
 
     // Test behavior for this status code against every endpoint
     // since not all business logic is shared for each.
     const endpointNames = Object.keys(endpointDataChecks)
-    // statusCodeTest.plan(endpointNames.length)
+    statusCodeTest.plan(endpointNames.length)
 
     endpointNames.forEach((endpointName) => {
       const checkHasTestData = endpointDataChecks[endpointName]
@@ -421,7 +307,7 @@ function whenAllAggregatorsSend(agent) {
  * Each type is added every test, even though not all endpoints are mocked.
  * This allows for verifying response handling for endpoint under test still
  * behaves correctly when other endpoints fail.
- * @param {*} agent The agent intance to add data to
+ * @param {*} agent The agent instance to add data to
  * @param {*} callback
  */
 function createTestData(agent, callback) {
