+++ third_party_manifest.json
@@ -1,5 +1,5 @@
 {
-  "lastUpdated": "Fri Oct 06 2023 17:25:04 GMT-0400 (Eastern Daylight Time)",
+  "lastUpdated": "Tue Oct 17 2023 15:36:00 GMT-0400 (Eastern Daylight Time)",
   "projectName": "New Relic Node Agent",
   "projectUrl": "https://github.com/newrelic/node-newrelic",
   "includeOptDeps": true,
